Relocated HTTP related functions.
