+++ src/gluon/services/jenkins/JenkinsService.ts
@@ -5,9 +5,9 @@ import * as https from "https";
 import * as _ from "lodash";
 import * as qs from "query-string";
 import * as util from "util";
-import {addAxiosLogger} from "../../util/shared/AxiosLogger";
+import {addAxiosLogger} from "../../../http/AxiosLogger";
+import {isSuccessCode} from "../../../http/Http";
 import {QMError} from "../../util/shared/Error";
-import {isSuccessCode} from "../../util/shared/Http";
 import {retryFunction} from "../../util/shared/RetryFunction";
 
 export class JenkinsService {
